check loop, createproductFavs and isadmin

+++ src/services/admin.services.js
@@ -30,11 +30,16 @@ const getUsers = async (values) => {
 };
 
 const deleteUser = async (userID) => {
-    try {
-        const response = await axios.delete(`${baseUrl}/${userID}`);
-        return response.data;
-    } catch (error) {
-        return error.message;
+
+    const drop = window.confirm('You are going to delete this profile. Are you sure?');
+
+    if (drop) {
+        try {
+            const response = await axios.delete(`${baseUrl}/${userID}`);
+            return response.data;
+        } catch (error) {
+            return error.message;
+        };
     };
 };
